nass-1260: useAutoUpdatingQuery thank u edwin wip"
