+++ packages/web/app/api/queries/index.js
@@ -11,4 +11,4 @@ export * from './usePatientProgramRegistration';
 export * from './useTranslations';
 export * from './useTranslationLanguages';
 export * from './useTelegramBotInfoQuery';
-export * from './useChangeDetectingQuery';
+export * from './useAddressHierarchyQuery';
